# docs/thv-registry-api/swagger.yaml
components:
  schemas:
    internal_api_v0.ErrorResponse:
      properties:
        error:
          type: string
      type: object
    internal_api_v0.RegistryInfoResponse:
      properties:
        last_updated:
          type: string
        source:
          type: string
        total_servers:
          type: integer
        version:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 'OAuth 2.0 Bearer token authentication. Format: "Bearer {token}"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    url: https://github.com/stacklok/toolhive
  description: |-
    API for accessing MCP server registry data and deployed server information
    This API provides endpoints to query the MCP (Model Context Protocol) server registry,
    get information about available servers, and check the status of deployed servers.

    Authentication is required by default. Use Bearer token authentication with a valid
    OAuth/OIDC access token. The /.well-known/oauth-protected-resource endpoint provides
    OAuth discovery metadata (RFC 9728).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ToolHive Registry API
  version: "0.1"
openapi: 3.1.0
paths:
...
  /api/v0/registry/info:
    get:
      deprecated: true
      description: Get registry metadata including version, last updated time, and
        total servers
      parameters:
      - description: Response format
        in: query
        name: format
        schema:
          default: toolhive
          enum:
          - toolhive
          - upstream
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_v0.RegistryInfoResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_v0.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_v0.ErrorResponse'
          description: Unauthorized
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_v0.ErrorResponse'
          description: Not Implemented
      security:
      - BearerAuth: []
      summary: Get registry information
      tags:
      - registry
...
  /readiness:
    get:
      deprecated: true
      description: Check if the registry API is ready to serve requests
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_api_v0.ErrorResponse'
          description: Service Unavailable
      summary: Readiness check
      tags:
      - system
...